♻️ Split compiler programs from pipeline
