+++ mlir/lib/Compiler/CMakeLists.txt
@@ -46,23 +46,60 @@ target_sources(
   MQTCompilerQDMIAdapter PUBLIC FILE_SET HEADERS BASE_DIRS ${MQT_MLIR_SOURCE_INCLUDE_DIR} FILES
                                 ${MQT_MLIR_SOURCE_INCLUDE_DIR}/mlir/Compiler/QDMIAdapter.h)
 
+# Build the compiler program model and OpenQASM-to-QCO path
+add_mlir_library(
+  MQTCompilerPrograms
+  PARTIAL_SOURCES_INTENDED
+  Programs.cpp
+  ADDITIONAL_HEADER_DIRS
+  ${MQT_MLIR_SOURCE_INCLUDE_DIR}/mlir/Compiler
+  LINK_LIBS
+  PUBLIC
+  LLVMSupport
+  MLIRArithDialect
+  MLIRBuiltinToLLVMIRTranslation
+  MLIRCBitDialect
+  MLIRControlFlowDialect
+  MLIRFuncDialect
+  MLIRIR
+  MLIRJeff
+  MLIRLLVMDialect
+  MLIRLLVMToLLVMIRTranslation
+  MLIRMathDialect
+  MLIRMemRefDialect
+  MLIRMQTDialect
+  MLIRParser
+  MLIRPass
+  MLIRQCDialect
+  MLIRQCOpenQASMTranslation
+  MLIRQCToQCO
+  MLIRQCODialect
+  MLIRQTensorDialect
+  MLIRSCFDialect
+  MLIRSupport
+  MLIRTensorDialect)
+
+mqt_mlir_target_use_project_options(MQTCompilerPrograms)
+
+target_sources(
+  MQTCompilerPrograms PUBLIC FILE_SET HEADERS BASE_DIRS ${MQT_MLIR_SOURCE_INCLUDE_DIR} FILES
+                             ${MQT_MLIR_SOURCE_INCLUDE_DIR}/mlir/Compiler/Programs.h)
+
 # Build the compiler pipeline library
 add_mlir_library(
   MQTCompilerPipeline
   PARTIAL_SOURCES_INTENDED
-  Programs.cpp
+  Pipeline.cpp
   TargetCompilation.cpp
   ADDITIONAL_HEADER_DIRS
   ${MQT_MLIR_SOURCE_INCLUDE_DIR}/mlir/Compiler
   LINK_LIBS
   PUBLIC
   LLVMBitWriter
-  MLIRCBitDialect
   MLIRPass
   MLIRTransforms
   MLIRTransformUtils
   MLIRMQTTransforms
-  MLIRQCToQCO
   MLIRQCOToQC
   MLIRQCOToJeff
   MLIRQCToQIRBase
@@ -70,22 +107,15 @@ add_mlir_library(
   MLIRQIRUtils
   MLIRQCOTransforms
   MLIRQCOpenQASMTranslation
-  MLIRParser
   MLIRJeffTranslation
   MLIRJeffToQCO
   MLIRTargetLLVMIRExport
-  MLIRBuiltinToLLVMIRTranslation
-  MLIRLLVMToLLVMIRTranslation
-  MLIRMathDialect
-  MLIRMQTDialect
+  MQTCompilerPrograms
   MQTCompilerTarget
   MQT::MLIRSupport)
 
 mqt_mlir_target_use_project_options(MQTCompilerPipeline)
 
-# collect header files
-file(GLOB_RECURSE COMPILER_HEADERS_SOURCE "${MQT_MLIR_SOURCE_INCLUDE_DIR}/mlir/Compiler/*.h")
-list(FILTER COMPILER_HEADERS_SOURCE EXCLUDE REGEX "/(QDMIAdapter|Target)\\.h$")
-
-target_sources(MQTCompilerPipeline PUBLIC FILE_SET HEADERS BASE_DIRS ${MQT_MLIR_SOURCE_INCLUDE_DIR}
-                                          FILES ${COMPILER_HEADERS_SOURCE})
+target_sources(
+  MQTCompilerPipeline PUBLIC FILE_SET HEADERS BASE_DIRS ${MQT_MLIR_SOURCE_INCLUDE_DIR} FILES
+                             ${MQT_MLIR_SOURCE_INCLUDE_DIR}/mlir/Compiler/TargetCompilation.h)
